refactor: extract complaint utilities
