+++ components/ErrorBoundary.tsx
@@ -14,6 +14,7 @@ import {
   Smartphone,
   Monitor
 } from 'lucide-react';
+import { LoadingSkeleton, EmptyState } from "@/components/complaints";
 
 interface ErrorBoundaryState {
   hasError: boolean;
@@ -379,27 +380,6 @@ export function LoadingCard({
     </Card>
   );
 }
-
-export function LoadingSkeleton({ 
-  rows = 3, 
-  className = '' 
-}: { 
-  rows?: number; 
-  className?: string; 
-}) {
-  return (
-    <div className={`space-y-3 ${className}`}>
-      {Array.from({ length: rows }).map((_, i) => (
-        <div key={i} className="animate-pulse flex space-x-4">
-          <div className="rounded-full bg-gray-200 dark:bg-gray-700 h-10 w-10 flex-shrink-0"></div>
-          <div className="flex-1 space-y-2 py-1">
-            <div className="h-4 bg-gray-200 dark:bg-gray-700 rounded w-3/4"></div>
-            <div className="h-4 bg-gray-200 dark:bg-gray-700 rounded w-1/2"></div>
-          </div>
-        </div>
-      ))}
-    </div>
-  );
 }
 
 export function LoadingTable({ 
@@ -431,35 +411,3 @@ export function LoadingTable({
   );
 }
 
-export function EmptyState({ 
-  icon: Icon = MessageCircle, 
-  title = 'ไม่มีข้อมูล', 
-  description = 'ไม่พบข้อมูลที่ต้องการ',
-  action,
-  className = '' 
-}: {
-  icon?: React.ComponentType<any>;
-  title?: string;
-  description?: string;
-  action?: ReactNode;
-  className?: string;
-}) {
-  return (
-    <div className={`text-center py-12 sm:py-16 ${className}`}>
-      <div className="w-16 h-16 bg-gray-100 dark:bg-gray-800 rounded-full flex items-center justify-center mx-auto mb-4">
-        <Icon className="w-8 h-8 text-gray-400" />
-      </div>
-      <h3 className="text-lg font-semibold text-gray-900 dark:text-white mb-2">
-        {title}
-      </h3>
-      <p className="text-gray-500 dark:text-gray-400 mb-6 max-w-md mx-auto">
-        {description}
-      </p>
-      {action && (
-        <div className="flex justify-center">
-          {action}
-        </div>
-      )}
-    </div>
-  );
-}
