+++ components/complaints/index.ts
@@ -1,0 +1,3 @@
+export { LoadingSkeleton } from './LoadingSkeleton';
+export { EmptyState } from './EmptyState';
+export { ComplaintDetailModal } from './ComplaintDetailModal';
